Update navigation and slider components
